refactor(luna-correlate): own KpPoint in kp.ts, derive AggMetric
Type-only polish (no behavior change):
- Relocate `KpPoint` from correlate.ts to kp.ts — correlate.ts is factor-agnostic
  (joins the generic FactorPoint); the kp parser/cache module owns the Kp type.
- Derive `AggMetric = keyof Omit<DailyAgg,"date">` so the metric set tracks the
  daily aggregates instead of a hand-maintained literal union.

83 tests pass; tsc --noEmit clean.

Security reviewed: pr-review-toolkit
